requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
anyio=3.1.0=py39hf3d152e_0
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=py39h27cfd23_1
async_generator=1.10=py_0
attrs=20.2.0=py_0
audioread=2.1.9=py39hf3d152e_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
brotlipy=0.7.0=py39h3811e60_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py39hf3d152e_0
cffi=1.14.5=py39he32792d_0
chardet=4.0.0=py39hf3d152e_1
cryptography=3.4.7=py39hbca0aa6_0
cycler=0.10.0=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py39h06a4308_0
expat=2.4.1=h2531618_2
ffmpeg=4.3.1=hca11adc_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.2=h36276a3_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py39hb070fc8_0
ipython=7.22.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.5.1=py_1
jedi=0.17.2=py39h06a4308_1
jinja2=2.11.2=py_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py39h06a4308_0
jupyter_server=1.8.0=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
kiwisolver=1.3.1=py39h2531618_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libllvm10=10.0.1=he513fc3_3
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0=py39h1bbdace_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=py39h27cfd23_0
matplotlib=3.3.4=py39h06a4308_0
matplotlib-base=3.3.4=py39h62a2d02_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py39h27cfd23_1
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.1=py39ha9443f7_2
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py39hf3d152e_3
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
notebook=6.4.0=py39h06a4308_0
numba=0.53.1=py39h56b8d98_1
numpy=1.20.2=py39h2d18471_0
numpy-base=1.20.2=py39hfae3a4d_0
olefile=0.46=py_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=py39h2531618_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py39h06a4308_1
parso=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py39he98fc37_0
pip=21.1.2=py39h06a4308_0
plotly=4.14.3=py_0
pooch=1.4.0=pyhd8ed1ab_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pynndescent=0.5.2=pyh44b312d_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.17.3=py39h27cfd23_0
pysocks=1.7.1=py39hf3d152e_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.9.5=h12debd9_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.9=1_cp39
pytz=2021.1=pyhd3eb1b0_0
pyzmq=20.0.0=py39h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
resampy=0.2.2=py_0
retrying=1.3.3=py_2
scikit-learn=0.24.2=py39ha9443f7_0
scipy=1.6.2=py39had2a1c9_1
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py39h06a4308_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py39hf3d152e_1
sqlite=3.35.4=hdfb4753_0
tbb=2020.2=h4bd325d_4
terminado=0.9.4=py39h06a4308_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.1=py39h27cfd23_0
traitlets=5.0.5=py_0
tzdata=2020f=h52ac0ba_0
umap-learn=0.5.1=py39hf3d152e_1
urllib3=1.26.5=pyhd8ed1ab_0
voila=0.2.10=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.57.0=py39hf3d152e_4
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py39h06a4308_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
